testing time dealer ui and stopping countdown during encounters, organize uis and loot choice uis.

+++ Assets/src/Util/SamplingUtil.cs
@@ -44,8 +44,8 @@ namespace Curry.Util
             //Sample probability variables
             float p;
             float sample;
-            int numToGet = numToSample;
-            int numLeftInRange = pool.Count;
+            float numToGet = numToSample;
+            float numLeftInRange = pool.Count;
             foreach (T v in pool)
             {
                 p = numToGet / numLeftInRange;
